Got 9b working.. finally
took awhile since I thought valid neighbours to be included had to be one more than the source so in the below example

514
203
149

The source is zero, so only the top neighbour is 1 more than it. I finally realized that it was any number that was not 9. 9s are the border.